+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added|Changed|Deprecated|Removed|Fixed|Security
 Nothing so far
 
+## 2.3.1 - 2020-04-23
+### Fixed
+- The `PredefinedKey::createJsonSchemaKey` now only takes the schema file.  All other information
+  will be obtained from the schema file:
+  + Key: the file name
+  + Value (default): the default values from the schema file
+  + FriendlyName: the `description` or `title` from the schema file with a fallback to the Key
+
 ## 2.3.0 - 2020-04-22
 ### Added
 - The bundle now supports keys based on a json schema.  The values will be configurable in the
@@ -18,8 +26,7 @@ Nothing so far
 
   ```php
   PredefinedKey::createJsonSchemaKey(
-      realpath(__DIR__ . '/../Resources/public/key-value-storage/this-is-the-key.schema.json'),
-      ['foo' => 'Foo']);
+      realpath(__DIR__ . '/../Resources/public/key-value-storage/this-is-the-key.schema.json'));
   ```
 
 ## 2.2.2 - 2019-09-10